% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNShiny.R
\name{KNNShiny}
\alias{KNNShiny}
\title{Launch 'KNNShiny' Interface}
\usage{
KNNShiny()
}
\value{
Nothing
}
\description{
KNNShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start KNNShiny

The interactive user interface is to provide an easy way for K-nearest neighbour(KNN) analaysis and downloading relevant plot.
}
\examples{
if(interactive()){
library(rmarkdown)
KNNShiny()
}
}
\keyword{KNNShiny}
